WiP: BDE overhaul
See https://github.com/nlesc-nano/nano-CAT/issues/1
